backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose
passlib[bcrypt]
python-dotenv
